requirements.txt
python-dotenv
requests
openai
langchain